[studio/frontend] sidenote -> aside
Making sidenotes inline caused many edge-case issues (especially with
pasting) and produced semantically incorrect HTML. Making them a block
element (with the existing HTML element `aside`) resolves these issues
but reduces the precision of where a sidenote can be placed. It also
should make it easier to translate Iris documents to other formats (as
asides can be placed in the main document flow).

Footnotes may be added in the future, although general popularity and
practicality of footnotes is dubious.